requirements.txt
accelerate==0.25.0
aiofiles==23.2.1
aiohttp==3.8.4
aiosignal==1.3.1
altair==5.2.0
annotated-types==0.6.0
anyio==3.7.1
async-timeout==4.0.2
attrs==23.1.0
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
contourpy==1.2.0
cycler==0.12.1
dataclasses-json==0.5.7
diskcache==5.6.3
exceptiongroup==1.2.0
fastapi==0.105.0
ffmpy==0.3.1
filelock==3.12.0
fonttools==4.46.0
frozenlist==1.3.3
fsspec==2023.5.0
gradio==4.9.0
gradio_client==0.7.2
greenlet==2.0.2
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
huggingface-hub==0.19.4
idna==3.4
importlib-resources==6.1.1
intel-openmp==2023.1.0
Jinja2==3.1.2
joblib==1.2.0
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
kiwisolver==1.4.5
langchain==0.0.350
langchain-community==0.0.2
langchain-core==0.1.0
langsmith==0.0.69
llama_cpp_python==0.2.20
markdown-it-py==2.2.0
MarkupSafe==2.1.3
marshmallow==3.19.0
marshmallow-enum==1.5.1
matplotlib==3.8.2
mdurl==0.1.2
mkl==2023.1.0
mkl-include==2023.1.0
multidict==6.0.4
mypy-extensions==1.0.0
nltk==3.8.1
numexpr==2.8.4
numpy==1.24.3
openapi-schema-pydantic==1.2.4
orjson==3.9.10
packaging==23.2
pandas==2.1.4
Pillow==9.5.0
psutil==5.9.5
pydantic==2.5.2
pydantic_core==2.14.5
pydub==0.25.1
Pygments==2.15.1
pyparsing==3.1.1
python-dateutil==2.8.2
python-multipart==0.0.6
pytz==2023.3.post1
PyYAML==6.0
referencing==0.32.0
regex==2023.5.5
requests==2.30.0
rich==13.3.5
rpds-py==0.13.2
safetensors==0.4.1
scikit-learn==1.2.2
scipy==1.10.1
semantic-version==2.10.0
sentence-transformers==2.2.2
sentencepiece==0.1.99
shellingham==1.5.4
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.14
starlette==0.27.0
tbb==2021.9.0
tenacity==8.2.2
threadpoolctl==3.1.0
tokenizers==0.15.0
tomlkit==0.12.0
toolz==0.12.0
torch==1.11.0
torchaudio==0.11.0
torchvision==0.12.0
tqdm==4.65.0
transformers==4.36.0
typer==0.9.0
typing-inspect==0.8.0
typing_extensions==4.9.0
tzdata==2023.3
urllib3==2.0.2
uvicorn==0.24.0.post1
websockets==11.0.3
yarl==1.9.2
